Commit Chingon de Ramon
En este commit chingon de Ramon podemos apreciar que ya se ha implementado la busqueda inteligentosa, asi mismo hemos visto que la seleccion del rectangulo a escanear de nuestra futurista y chingonsisima aplicacion ahora se puede hacer de manera manual. WOOOOWW que increible desarrollador esclavo es Jose Ramon
