+++ requirements.txt
@@ -64,6 +64,7 @@ pycparser==2.22
 pydyf==0.11.0
 pyparsing==3.2.3
 pyphen==0.17.2
+PyPDF2==3.0.1
 PyQt5==5.15.11
 PyQt5-Qt5==5.15.16
 PyQt5_sip==12.17.0
@@ -81,6 +82,7 @@ scikit-image==0.25.2
 scipy==1.15.2
 setuptools==78.1.0
 shapely==2.1.0
+spacy==3.6.0
 sympy==1.13.1
 tesseract==0.1.3
 tifffile==2025.3.30
