+++ templates/mutable/provider.tmpl
@@ -1,0 +1,20 @@
+import 'dart:async';
+
+class ${upperName}Provider {
+  Future<void> loadAsync(String token) async {
+    /// write from keystore/keychain
+    await Future.delayed(Duration(seconds: 2));
+  }
+
+  Future<void> saveAsync(String token) async {
+    /// write from keystore/keychain
+    await Future.delayed(Duration(seconds: 2));
+  }
+
+  void test(bool isError) {
+    if (isError == true){
+      throw Exception('manual error');
+    }
+  }
+}
+
