+++ templates/simple/model.tmpl
@@ -1,0 +1,13 @@
+import 'package:equatable/equatable.dart';
+
+/// use https://marketplace.visualstudio.com/items?itemName=BendixMa.dart-data-class-generator
+class ${upperName}Model extends Equatable {
+  final int id;
+  final String name;
+
+  ${upperName}Model(this.id, this.name);
+
+  @override
+  List<Object> get props => [id, name];
+
+}
