+++ templates/simple/repository.tmpl
@@ -1,0 +1,11 @@
+import 'package:${appName}${relative}/index.dart';
+
+class ${upperName}Repository {
+  final ${upperName}Provider _${privateName}Provider = ${upperName}Provider();
+
+  ${upperName}Repository();
+
+  void test(bool isError) {
+    _${privateName}Provider.test(isError);
+  }
+}
