+++ ventanas/ventanaVisualizacion.py
@@ -1,6 +1,6 @@
 from PyQt5.QtWidgets import QDialog, QVBoxLayout, QHBoxLayout, QGraphicsView, QGraphicsScene, QPushButton, QSlider, QLabel
 from PyQt5.QtGui import QPixmap, QPen, QCursor, QImage
-from PyQt5.QtCore import Qt
+from PyQt5.QtCore import Qt, QLineF
 import cv2
 import numpy as np
 
@@ -12,9 +12,10 @@ class VentanaVisualizacion(QDialog):
         self.imagen_actual = imagen
         self.pixmap = QPixmap.fromImage(imagen)
         self.puntos = []  # Lista para almacenar los puntos seleccionados
+        self.linea_guia = None  # Línea guía temporal
 
         # Establecer un tamaño fijo para la ventana
-        self.resize(800, 600)  # Tamaño fijo de la ventana (800x600)
+        self.resize(800, 600)
 
         # Configurar la escena y la vista
         self.scene = QGraphicsScene(self)
@@ -33,6 +34,12 @@ class VentanaVisualizacion(QDialog):
 
         # Añadir controles de ajustes
         self.crear_controles_ajustes()
+
+        # Añadir texto flotante
+        self.texto_flotante = QLabel("Presiona E para colocar una esquina", self)
+        self.texto_flotante.setStyleSheet("color: white; background-color: black; padding: 5px;")
+        self.texto_flotante.setAlignment(Qt.AlignCenter)
+        self.texto_flotante.setGeometry(10, 10, 300, 30)  # Posición y tamaño del texto flotante
 
         # Habilitar zoom y arrastre
         self.view.setDragMode(QGraphicsView.ScrollHandDrag)
@@ -62,11 +69,6 @@ class VentanaVisualizacion(QDialog):
         layout_controles.addWidget(QLabel("Contraste"))
         layout_controles.addWidget(self.slider_contraste)
 
-        # Botón para pasar a la selección de esquinas
-        boton_seleccionar_esquinas = QPushButton("Seleccionar Esquinas")
-        boton_seleccionar_esquinas.clicked.connect(self.iniciar_seleccion_esquinas)
-        layout_controles.addWidget(boton_seleccionar_esquinas)
-
         # Añadir los controles al diseño principal
         self.layout_principal.addLayout(layout_controles)
 
@@ -87,25 +89,37 @@ class VentanaVisualizacion(QDialog):
         """Inicia la fase de selección de esquinas."""
         self.slider_brillo.setEnabled(False)
         self.slider_contraste.setEnabled(False)
-        self.view.setDragMode(QGraphicsView.NoDrag)
-        self.scene.mousePressEvent = self.mouse_press_event
+        self.view.setDragMode(QGraphicsView.NoDrag)  # Deshabilitar el modo de arrastre
+        self.view.viewport().setCursor(QCursor(Qt.CrossCursor))  # Cambiar el cursor a una cruceta
+        self.scene.mousePressEvent = self.mouse_press_event  # Conectar el evento de clic a la escena
 
     def mouse_press_event(self, event):
         """Registra los puntos seleccionados al hacer clic en la imagen."""
-        if len(self.puntos) < 4:  # Limitar a 4 puntos
-            punto = self.view.mapToScene(event.scenePos().toPoint())
-            self.puntos.append(punto)
-            self.dibujar_punto(punto)
+        # Convertir la posición del clic a coordenadas de la escena
+        punto = event.scenePos()
+        self.puntos.append(punto)  # Agregar el punto a la lista
+        self.dibujar_punto(punto)  # Dibujar el punto rojo
 
-            # Si se seleccionan 4 puntos, cerrar la ventana
-            if len(self.puntos) == 4:
-                self.close()
+        # Dibujar la línea guía o cerrar el rectángulo
+        if len(self.puntos) > 1:
+            self.dibujar_linea_permanente(self.puntos[-2], self.puntos[-1])  # Línea entre los últimos dos puntos
+
+        if len(self.puntos) == 4:
+            # Cerrar el rectángulo
+            self.dibujar_linea_permanente(self.puntos[-1], self.puntos[0])  # Línea entre el último y el primer punto
+            self.close()  # Cerrar la ventana una vez seleccionados los 4 puntos
 
     def dibujar_punto(self, punto):
         """Dibuja un punto en la escena."""
         pen = QPen(Qt.red)
         pen.setWidth(5)
         self.scene.addEllipse(punto.x() - 2, punto.y() - 2, 4, 4, pen)
+
+    def dibujar_linea_permanente(self, punto_inicio, punto_fin):
+        """Dibuja una línea permanente entre dos puntos."""
+        pen = QPen(Qt.yellow)
+        pen.setWidth(2)
+        self.scene.addLine(QLineF(punto_inicio, punto_fin), pen)
 
     def actualizar_imagen(self, imagen_cv):
         """Actualiza la imagen mostrada en la vista."""
@@ -118,6 +132,23 @@ class VentanaVisualizacion(QDialog):
         """Controla el zoom con el scroll del mouse."""
         factor = 1.25 if event.angleDelta().y() > 0 else 0.8
         self.view.scale(factor, factor)
+
+    def keyPressEvent(self, event):
+        """Captura la tecla presionada para colocar una esquina."""
+        if event.key() == Qt.Key_E:
+            # Obtener la posición actual del cursor en la escena
+            cursor_pos = self.view.mapToScene(self.view.viewport().mapFromGlobal(QCursor.pos()))
+            self.puntos.append(cursor_pos)  # Agregar el punto a la lista
+            self.dibujar_punto(cursor_pos)  # Dibujar el punto rojo
+
+            # Dibujar la línea guía o cerrar el rectángulo
+            if len(self.puntos) > 1:
+                self.dibujar_linea_permanente(self.puntos[-2], self.puntos[-1])  # Línea entre los últimos dos puntos
+
+            if len(self.puntos) == 4:
+                # Cerrar el rectángulo
+                self.dibujar_linea_permanente(self.puntos[-1], self.puntos[0])  # Línea entre el último y el primer punto
+                self.close()  # Cerrar la ventana una vez seleccionados los 4 puntos
 
     @staticmethod
     def qimage_a_cv2(qimage):
@@ -137,12 +168,3 @@ class VentanaVisualizacion(QDialog):
         bytes_per_line = 3 * width
         qimage = QImage(cv_image.data, width, height, bytes_per_line, QImage.Format.Format_RGB888)
         return qimage
-
-    def obtener_puntos(self):
-        """Obtiene los puntos seleccionados como tuplas de enteros."""
-        puntos = [(int(p.x()), int(p.y())) for p in self.puntos]
-        pts = np.array(puntos, dtype="float32")
-        rect = cv2.boundingRect(pts)
-        x, y, w, h = rect
-        recorte = self.qimage_a_cv2(self.imagen_original)[y:y+h, x:x+w]
-        return puntos
